+++ lib/adyen-ruby-api-library.rb
@@ -14,3 +14,4 @@ require_relative "adyen/services/bin_lookup"
 require_relative "adyen/services/balance_platform"
 require_relative "adyen/hash_with_accessors"
 require_relative "adyen/utils/hmac_validator"
+require_relative "adyen/services/transfers"
