+++ lib/adyen/services/balance_platform.rb
@@ -132,21 +132,5 @@ module Adyen
 
       @client.call_adyen_api(@service, action, {}, {}, @version)
     end
-    
-    def get_transactions(request_params)
-      action = { method: 'get', url: "transactions" }
-
-      @client.call_adyen_api(@service, action, request_params, {}, 2)
-    end
-
-    def get_transaction(transaction_id)
-      action = { method: 'get', url: "transactions/" + transaction_id }
-
-      @client.call_adyen_api(@service, action, {}, {}, 2)
-    end
-
-    def create_transfer_request(request)
-      @client.call_adyen_api(@service, "transfers", request, {}, 2)
-    end
   end
 end
