+++ spec/balance_platform_spec.rb
@@ -527,85 +527,5 @@ RSpec.describe Adyen::BalancePlatform, service: "Balance Platform service" do
     end
   end
 
-  context "transaction" do
-    it "retrieves transactions" do
-      balance_account_id = "BA32272223222B5FD6CD2FNXB"
-      response_body = json_from_file("mocks/responses/BalancePlatform/get_transactions.json")
-
-      url = client.service_url("BalancePlatform", "transactions", "2")
-
-      request_params = {
-        "balanceAccountId" => balance_account_id,
-        "createdSince" => "2021-01-01T15:07:40Z",
-        "createdUntil" => "2022-12-31T15:07:40Z",
-        "limit" => "100"
-      }
-
-      WebMock.stub_request(:get, url).
-        with(
-          headers: {
-            "x-api-key" => client.api_key
-          },
-          query: request_params
-        ).
-        to_return(
-          body: response_body
-        )
-
-      result = client.balance_platform.get_transactions(request_params)
-      response_hash = result.response
-
-      expect(result.status).
-        to eq(200)
-      expect(response_hash).
-        to eq(JSON.parse(response_body))
-      expect(response_hash).
-        to be_a Adyen::HashWithAccessors
-      expect(response_hash).
-        to be_a_kind_of Hash
-    end
-
-    it "retrieves information about a single transaction" do
-      transaction_id = "3JERI55U58GRGWCK"
-      response_body = json_from_file("mocks/responses/BalancePlatform/get_transaction.json")
-
-      url = client.service_url("BalancePlatform", "transactions/#{transaction_id}", "2")
-
-      WebMock.stub_request(:get, url).
-        with(
-          headers: {
-            "x-api-key" => client.api_key
-          }
-        ).
-        to_return(
-          body: response_body
-        )
-
-      result = client.balance_platform.get_transaction(transaction_id)
-      response_hash = result.response
-
-      expect(result.status).
-        to eq(200)
-      expect(response_hash).
-        to eq(JSON.parse(response_body))
-      expect(response_hash).
-        to be_a Adyen::HashWithAccessors
-      expect(response_hash).
-        to be_a_kind_of Hash
-
-    end
-  end
-
-  context "transfers" do
-    xit "creates a transfer" do
-      request_body = JSON.parse(json_from_file("mocks/requests/BalancePlatform/create_transfer_request.json"))
-
-      result = client.balance_platform.create_transfer_request(request_body)
-      response_hash = result.response
-      puts response_hash
-      # TODO: Find a balance account with capabilites and balance > 0
-    end
-  end
-
   generate_tests(client, "BalancePlatform", test_sets, client.balance_platform)
 end
